Use utility to create executor and fix executor shutdown
